generalise WflowMgrKafkaBlder and WflowMgrBldr
This commit makes WorkflowManagerBaseBuilder a parent class for WorkflowManagerKafkaBuilder and WorkflowManagerBuilder. While this implementation reduces boilerplate, it also opens up prospects for easy switching of workflow type (from Kafka based to Curator based and vice-versa) as both the variants of the builders inherit from a base class which opens up prospects for chaining the `withCurator` or `withKafka` methods, while building the Manager depending upon the type of workflow being used.